services/pathway/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
jinja2>=3.1.0
python-multipart>=0.0.6
openai>=1.3.0
neo4j>=5.13.0
PyMuPDF>=1.23.0
python-dotenv>=1.0.0
pandas>=2.0.0
pydantic>=2.0.0
openpyxl>=3.1.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
faiss-cpu>=1.8.0
rapidfuzz>=3.0.0
scikit-learn>=1.3.0
